refactor: move showNotification to its own file
