+++ src/game-logic/reducers/index.js
@@ -94,7 +94,6 @@ import { itemType } from '../../enums'
 
 import { applyChanceEvent } from './helpers'
 
-import { showNotification } from './common'
 import { addItemToInventory } from './addItemToInventory'
 import { decrementItemFromInventory } from './decrementItemFromInventory'
 import {
@@ -104,9 +103,10 @@ import {
   processWeather,
   processSprinklers,
 } from './field'
+import { showNotification } from './showNotification'
 
-export * from './common'
 export * from './field'
+export * from './showNotification'
 export * from './decrementItemFromInventory'
 export * from './addItemToInventory'
 export * from './incrementPlotContentAge'
